third_party/WebKit/LayoutTests/css3/blending/blend-mode-property-expected.txt

Test the parsing of the mix-blend-mode property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'normal'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'multiply'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'screen'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'overlay'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'darken'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'lighten'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'color-dodge'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'color-burn'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'hard-light'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'soft-light'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'difference'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'exclusion'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'hue'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'saturation'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'color'
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'luminosity'
PASS successfullyParsed is true

TEST COMPLETE
